Add status for cached searches
LLM getting into loop and not answering (even though the answer is there)
Tried Locally Perplexity.AI works amazing for summary or information kind of querries. Its way Faster as it avoids steps of navigating link to get info. We need to have server working with it.
Increased the viewport to see more of webpage to allow screenshot to have more information
Updated Prompt to present detailed summary. Emphasised that the user is blind and this software helps then navigate. This results in more data coming from screenshot